+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 - EntityPanel.Builder.icon() added.
 - EntityPanel, caption, description and icon moved to Config, related changes.
 - EntityPanel.disposeEditDialogOnEscape moved to Config.
+- EntityPanel.updateEditPanelState() refactored, ShowHiddenEditPanel now brings window to front if available, minor related changes.
 
 ## 0.17.37
 ### is.codion.common.core